Refactor Query helper into own Extension class
